chore: move to a package with a more consistent name
