+++ docker_client.go
@@ -10,8 +10,8 @@ import (
 	"github.com/docker/docker/api/types/registry"
 	"github.com/docker/docker/client"
 
+	"github.com/testcontainers/testcontainers-go/internal/core"
 	"github.com/testcontainers/testcontainers-go/internal/testcontainersdocker"
-	"github.com/testcontainers/testcontainers-go/internal/testcontainerssession"
 )
 
 // DockerClient is a wrapper around the docker client that is used by testcontainers-go.
@@ -68,8 +68,8 @@ func (c *DockerClient) Info(ctx context.Context) (types.Info, error) {
 		dockerInfo.OperatingSystem, dockerInfo.MemTotal/1024/1024,
 		testcontainersdocker.ExtractDockerHost(ctx),
 		testcontainersdocker.ExtractDockerSocket(ctx),
-		testcontainerssession.SessionID(),
-		testcontainerssession.ProcessID(),
+		core.SessionID(),
+		core.ProcessID(),
 	)
 
 	return dockerInfo, nil
