+++ generic.go
@@ -6,8 +6,8 @@ import (
 	"fmt"
 	"sync"
 
+	"github.com/testcontainers/testcontainers-go/internal/core"
 	"github.com/testcontainers/testcontainers-go/internal/testcontainersdocker"
-	"github.com/testcontainers/testcontainers-go/internal/testcontainerssession"
 )
 
 var (
@@ -94,5 +94,5 @@ type GenericProvider interface {
 
 // GenericLabels returns a map of labels that can be used to identify containers created by this library
 func GenericLabels() map[string]string {
-	return testcontainersdocker.DefaultLabels(testcontainerssession.SessionID())
+	return testcontainersdocker.DefaultLabels(core.SessionID())
 }
