+++ internal/testcontainersdocker/client.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/testcontainers/testcontainers-go/internal"
 	"github.com/testcontainers/testcontainers-go/internal/config"
-	"github.com/testcontainers/testcontainers-go/internal/testcontainerssession"
+	"github.com/testcontainers/testcontainers-go/internal/core"
 )
 
 // NewClient returns a new docker client extracting the docker host from the different alternatives
@@ -33,8 +33,8 @@ func NewClient(ctx context.Context, ops ...client.Opt) (*client.Client, error) {
 
 	opts = append(opts, client.WithHTTPHeaders(
 		map[string]string{
-			"x-tc-pp":    testcontainerssession.ProjectPath(),
-			"x-tc-sid":   testcontainerssession.SessionID(),
+			"x-tc-pp":    core.ProjectPath(),
+			"x-tc-sid":   core.SessionID(),
 			"User-Agent": "tc-go/" + internal.Version,
 		}),
 	)
